try with waitFor has text
